Fix CI: resolve ktlint/detekt findings in the voting module
static_analysis_ktlint and static_analysis_detekt were failing
pre-existing on this branch (54 ktlint violations, plus detekt
findings confined to VotingSdkImpl.kt/VotingSdkMappers.kt/
VotingModels.kt).

- ktlint: ran `./gradlew ktlintFormat`, which was entirely mechanical
  (import ordering, call-argument wrapping, no logic changes) across
  VotingSdkImpl.kt, VotingSdkMappers.kt, VotingSdk.kt, VotingModels.kt,
  and the two Voting*Test.kt files.

- detekt TooManyFunctions (VotingSdkMappers.kt: 28 functions, threshold
  11): split the file by domain into VotingSdkNoteMappers.kt,
  VotingSdkRoundMappers.kt, VotingSdkVoteMappers.kt, and
  VotingSdkDelegationMappers.kt. Pure file split, no behavior changes.
  While splitting, also replaced several inline fully-qualified type
  references with plain imports (using an `as`-alias for the one
  genuine name collision, `VotingTxHashLookup`, which exists in both
  the internal and public model packages) -- the FQNs weren't needed
  for the other mappers and were pushing lines past detekt's
  MaxLineLength.

- detekt CyclomaticComplexMethod (VotingModels.kt:141,
  VotingCommitmentResult.equals(), complexity 18 vs threshold 15):
  split the single long `&&`-chained comparison into three
  private per-field-group helpers (scalarFieldsMatch,
  byteArrayFieldsMatch, collectionFieldsMatch) so `equals()` itself
  stays simple. Same equality semantics, just organized into readable
  chunks small enough to individually satisfy the complexity budget.

- detekt MaxLineLength: wrapped several long single-expression
  function bodies/signatures in VotingSdkImpl.kt, VotingSdkMappers.kt
  (now split), and VotingSdk.kt (including one KDoc comment).

Verified: `make ktlint`, `make detekt`, `:sdk-lib:compileDebugKotlin`,
`:sdk-lib:compileDebugUnitTestKotlin`,
`:sdk-lib:compileDebugAndroidTestKotlin`, and
`:sdk-lib:testDebugUnitTest --tests "*Voting*"` all pass clean.
